+++ app/src/main/java/kr/co/company/capstone/createGoal/first/CreateGoalFirstPageContract.kt
@@ -1,0 +1,4 @@
+package kr.co.company.capstone.createGoal.first
+
+class CreateGoalFirstPageContract {
+}
